[ADD] add function get payroll by employerId

+++ Common/DTO/PayrollFiltersRequest.cs
@@ -11,8 +11,8 @@ namespace Common.DTO
         public int EmployerId { get; set; }
         public string NameEmployer { get; set; }
         public string EmailEmployer { get; set; }
-        public int Month { get; set; }
-        public int Year { get; set; }
+        public Int16 Month { get; set; }
+        public Int16 Year { get; set; }
         public byte Status { get; set; }
         public int Limit { get; set; }
     }
